requirements.txt
flask
gunicorn
python-dotenv
flask-wtf
pymongo
flask-security
email_validator
passlib
requests